Structure and style fixes
