+++ src/Tools/Common/ProcessNativeMethods/ProcessNativeMethods.cs
@@ -1,0 +1,126 @@
+// using Microsoft.Diagnostics.NETCore.Client;
+// using Microsoft.Tools.Common;
+using System;
+// using System.CommandLine;
+// using System.CommandLine.Invocation;
+// using System.CommandLine.IO;
+// using System.Collections.Generic;
+// using System.Diagnostics;
+// using System.Linq;
+using System.Runtime.InteropServices;
+// using System.Text;
+// using Process = System.Diagnostics.Process;
+// using System.IO;
+// using System.ComponentModel;
+// using System.Threading.Tasks;
+// using Microsoft.Diagnostics.Tools.Trace.CommandLine;
+// using System.CommandLine.Binding;
+
+namespace Microsoft.Internal.Common.Utils
+{
+    internal static class ProcessNativeMethods
+    {
+        public const int ProcessBasicInformation = 0;
+        public const int ProcessWow64Information = 26;
+
+        [DllImport("kernel32.dll", SetLastError = true)]
+        [return: MarshalAs(UnmanagedType.Bool)]
+        public static extern bool ReadProcessMemory(
+            IntPtr hProcess,
+            IntPtr lpBaseAddress,
+            [Out] byte[] lpBuffer,
+            IntPtr dwSize,
+            ref IntPtr lpNumberOfBytesRead);
+
+        [DllImport("kernel32.dll", SetLastError = true)]
+        [return: MarshalAs(UnmanagedType.Bool)]
+        public static extern bool ReadProcessMemory(
+            IntPtr hProcess,
+            IntPtr lpBaseAddress,
+            [Out] byte[] lpBuffer,
+            IntPtr dwSize,
+            IntPtr lpNumberOfBytesRead);
+
+        [DllImport("kernel32.dll", SetLastError = true)]
+        [return: MarshalAs(UnmanagedType.Bool)]
+        public static extern bool ReadProcessMemory(
+            IntPtr hProcess,
+            IntPtr lpBaseAddress,
+            out IntPtr lpPtr,
+            IntPtr dwSize,
+            ref IntPtr lpNumberOfBytesRead);
+
+        [DllImport("kernel32.dll", SetLastError = true)]
+        [return: MarshalAs(UnmanagedType.Bool)]
+        public static extern bool ReadProcessMemory(
+            IntPtr hProcess,
+            IntPtr lpBaseAddress,
+            ref UNICODE_STRING lpBuffer,
+            IntPtr dwSize,
+            IntPtr lpNumberOfBytesRead);
+
+        [DllImport("kernel32.dll", SetLastError = true)]
+        [return: MarshalAs(UnmanagedType.Bool)]
+        public static extern bool ReadProcessMemory(
+            IntPtr hProcess,
+            IntPtr lpBaseAddress,
+            ref UNICODE_STRING_32 lpBuffer,
+            IntPtr dwSize,
+            IntPtr lpNumberOfBytesRead);
+
+        [DllImport("advapi32.dll", SetLastError = true)]
+        [return: MarshalAs(UnmanagedType.Bool)]
+        public static extern bool OpenProcessToken(
+            IntPtr hProcess,
+            UInt32 dwDesiredAccess,
+            out IntPtr processToken);
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct UNICODE_STRING
+        {
+            public ushort Length;
+            public ushort MaximumLength;
+            public IntPtr Buffer;
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct UNICODE_STRING_32
+        {
+            public ushort Length;
+            public ushort MaximumLength;
+            public int Buffer;
+        }
+
+        [DllImport("kernel32.dll")]
+        [return: MarshalAs(UnmanagedType.Bool)]
+        public static extern bool IsWow64Process(IntPtr hProcess, [MarshalAs(UnmanagedType.Bool)]out bool wow64Process);
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct ProcessInformation
+        {
+            // These members must match PROCESS_BASIC_INFORMATION
+            internal IntPtr Reserved1;
+            internal IntPtr PebBaseAddress;
+            internal IntPtr Reserved2_0;
+            internal IntPtr Reserved2_1;
+            internal IntPtr UniqueProcessId;
+            internal IntPtr InheritedFromUniqueProcessId;
+        }
+
+        [DllImport("ntdll.dll")]
+        public static extern int NtQueryInformationProcess(
+                IntPtr processHandle,
+                int processInformationClass,
+                ref ProcessInformation processInformation,
+                int processInformationLength,
+                out int returnLength);
+
+        [DllImport("ntdll.dll", SetLastError = true)]
+        public static extern int NtQueryInformationProcess(
+            IntPtr processHandle,
+            int processInformationClass,
+            ref IntPtr processInformation,
+            int processInformationLength,
+            ref int returnLength);
+    }
+}
