+++ src/Tools/dotnet-counters/dotnet-counters.csproj
@@ -12,10 +12,7 @@
 
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)..\dotnet-trace\Extensions.cs" Link="Extensions.cs" />
-    <Compile Include="..\Common\CommandExtensions.cs" Link="CommandExtensions.cs" />
-    <Compile Include="..\Common\Commands\ProcessStatus.cs" Link="ProcessStatus.cs" />
-    <Compile Include="..\Common\ReversedServerHelpers\ReversedServerHelpers.cs" Link="ReversedServerHelpers.cs" />
-    <Compile Include="..\Common\Commands\Utils.cs" Link="Utils.cs" />
+    <Compile Include="..\Common\**\*.cs" />
   </ItemGroup>
 
   <ItemGroup>
