+++ src/Tools/dotnet-trace/dotnet-trace.csproj
@@ -20,10 +20,12 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="..\Common\CommandExtensions.cs" Link="CommandExtensions.cs" />
+    <!-- <Compile Include="..\Common\CommandExtensions.cs" Link="CommandExtensions.cs" />
     <Compile Include="..\Common\Commands\ProcessStatus.cs" Link="ProcessStatus.cs" />
     <Compile Include="..\Common\Commands\Utils.cs" Link="Utils.cs" />
 	  <Compile Include="..\Common\ReversedServerHelpers\ReversedServerHelpers.cs" Link="ReversedServerHelpers.cs" />
+    <Compile Include="..\Common\WindowsProcessExtension\WindowsProcessExtension.cs" Link="WindowsProcessExtension.cs" /> -->
+    <Compile Include="..\Common\**\*.cs" />
   </ItemGroup>
 
   <ItemGroup>
